2025.04.24 대공사
- PlayerUI, PlayerInput 클래스 삭제처리, 해당 클래스들 안의 배틀 관련 기능들을 BattleManager 클래스 내 통합
