+++ NotepadKnights/MainMenuModule.cs
@@ -42,7 +42,7 @@ namespace NotepadKnights
                 if (answer == 1)
                 {
                     Console.WriteLine("\n");
-                    Program.playerStatus.SetName(name);
+                    //Program.playerStatus.SetName(name);
                     break;
                 }
                 else if (answer == 2)
@@ -76,7 +76,7 @@ namespace NotepadKnights
                         Console.WriteLine("전사로 시작합니다.");
                         // 잠깐 테스트를 위해 변경했어요.
                         // Program.player.Job = "전사";
-                        Program.playerStatus.SetJob("전사");
+                        // Program.playerStatus.SetJob("전사");
                         Program.playerStatus.InitializePlayer();
                         break;
                     }
@@ -104,7 +104,7 @@ namespace NotepadKnights
                         Console.WriteLine("도적으로 시작합니다.");
                         // 잠깐 테스트를 위해 변경했어요.
                         // player.Job = "도적";
-                        Program.playerStatus.SetJob("도적");
+                        //Program.playerStatus.SetJob("도적");
                         Program.playerStatus.InitializePlayer();
                         break;
                     }
