+++ NotepadKnights/Player/PlayerStatus.cs
@@ -41,60 +41,5 @@ namespace NotepadKnights
         {
             return (Name, Job, Level, Hp, MaxHp, Mp, Exp, MaxExp, Gold, Attack, Defense);
         }
-        // 혹시 몰라서 넣음 , 이름 변경
-        public void SetName(string name)
-        {
-            Name = name;
-        }
-        // 직업 변경
-        public void SetJob(string job)
-        {
-            Job = job;
-        }
-        // 체력 변경
-        public void SetHp(int hp)
-        {
-            Hp = hp;
-        }
-        // 레벨 변경
-        public void SetLevel(int level)
-        {
-            Level = level;
-        }
-
-        // 현재 경험치 변경
-        public void SetExp(int exp)
-        {
-            Exp = exp;
-        }
-        // 최대 경험치 변경
-        public void SeMaxExp(int maxExp)
-        {
-            MaxExp = maxExp;
-        }
-        // 데미지 변경
-        public void SetAttack(float attack)
-        {
-            Attack = attack;
-        }
-        // 방어력 변경
-        public void SetDefense(int defense)
-        {
-            Defense = defense;
-        }
-        // 골드 변경
-        public void SetGold(int gold)
-        {
-            Gold = gold;
-        }
-        public void SetIsAttack(bool isAttack)
-        {
-            IsAttack = isAttack;
-        }
-        // 죽인 몬스터의 수 변경
-        public void SetKilledMonsterCount(int killedMonsterCount)
-        {
-            KilledMonsterCount = killedMonsterCount;
-        }
     }
 }
