+++ NotepadKnights/Program.cs
@@ -3,12 +3,11 @@
     //18조 TextRPG [Notepad Knights] 협업 공간입니다.
     internal class Program
     {
-
-        public static PlayerUI playerUI;
        // public static Player player;
         public static PlayerStatus playerStatus;
         public static MonsterFactory monsterFactory;
 
+        public static BattleManager battleManager = new BattleManager();
         public static Player player = new Player();
         public static MainMenuModule mainMenu = new MainMenuModule(player);
         public static Healing healing = new Healing();
@@ -18,7 +17,6 @@
         {
                 player = new Player();
                 playerStatus = new PlayerStatus();
-                playerUI = new PlayerUI();
                 monsterFactory = new MonsterFactory();
                 quest = new QuestUI(playerStatus);
                 mainMenu.Intro();
@@ -37,7 +35,7 @@
                         break;
                     case 4:
                         //전투
-                        playerUI.ShowBattleMenu();
+                        battleManager.ShowBattleMenu();
                         break;
                     case 5:
                         healing.IntoHealing();
